@@ -3,6 +3,7 @@
   "private": true,
   "description": "The Minimal House Remix Stack to setup Remix with minimal typsescript, tailwind, linting, and formatting. Worry about storage, testing, or deployment later.",
   "license": "MIT",
+  "type": "module",
   "sideEffects": false,
   "scripts": {
     "build": "run-s build:*",
@@ -11,7 +12,7 @@
     "dev": "run-p dev:*",
     "dev:css": "npm run generate:css -- --watch",
     "dev:remix": "remix dev",
-    "generate:css": "npx tailwindcss -i ./app/styles/app.css -o ./app/tailwind.css",
+    "generate:css": "npx @tailwindcss/cli -i ./app/styles/app.css -o ./app/tailwind.css",
     "lint": "eslint .",
     "lint:fix": "eslint --fix .",
     "start": "remix-serve build",
@@ -19,26 +20,30 @@
     "package": "create-registry-image --latest --push micrantha/web"
   },
   "dependencies": {
-    "@remix-run/node": "^1.3.5",
-    "@remix-run/react": "^1.3.5",
-    "@remix-run/serve": "^1.3.5",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "@remix-run/node": "2.17.2",
+    "@remix-run/react": "2.17.2",
+    "@remix-run/serve": "2.17.2",
+    "react": "19.2.0",
+    "react-dom": "19.2.0",
     "react-websitecarbon-badge": "^1.0.5"
   },
   "devDependencies": {
-    "@remix-run/dev": "^1.3.5",
-    "@remix-run/eslint-config": "^1.3.5",
+    "@eslint/compat": "^2.0.0",
+    "@remix-run/dev": "2.17.2",
+    "@remix-run/eslint-config": "2.17.2",
+    "@tailwindcss/cli": "^4.1.17",
     "@types/minimatch": "^6.0.0",
-    "@types/react": "^18.2.14",
-    "@types/react-dom": "^18.2.6",
-    "eslint": "^8.11.0",
-    "eslint-plugin-prettier": "^5.0.0",
+    "@types/react": "19.2.5",
+    "@types/react-dom": "19.2.3",
+    "eslint": "9.39.1",
+    "eslint-plugin-prettier": "^5.5.4",
+    "eslint-plugin-react": "^7.37.5",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.0.0",
-    "prettier-plugin-tailwindcss": "^0.5.4",
-    "tailwindcss": "^3.0.24",
-    "typescript": "^5.1.6"
+    "prettier-plugin-tailwindcss": "0.7.1",
+    "tailwindcss": "4.1.17",
+    "typescript": "^5.1.6",
+    "typescript-eslint": "^8.46.4"
   },
   "engines": {
     "node": ">=14"
